requirements.txt
streamlit>=1.28.0
openai>=1.3.0
anthropic>=0.25.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0 